data_processor/requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
PyYAML>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0